Move exec system call stuff to a separate module
To be used by the docker module in subsequent commits
