+++ modules/exec/src/main/java/coursier/exec/ErrnoException.java
@@ -1,9 +1,8 @@
-package coursier.jvm;
+package coursier.exec;
 
 import com.oracle.svm.core.posix.headers.Errno;
 import org.graalvm.nativeimage.c.type.CTypeConversion;
 
-@Deprecated
 public class ErrnoException extends Exception {
 
     private final int errno0;
