+++ modules/exec/src/main/java/coursier/exec/ExecveGraalvm.java
@@ -1,4 +1,4 @@
-package coursier.jvm;
+package coursier.exec;
 
 import java.io.FileNotFoundException;
 
@@ -9,8 +9,7 @@ import org.graalvm.nativeimage.c.type.CTypeConversion;
 import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.Platforms;
 
-@Deprecated
-@TargetClass(className = "coursier.jvm.Execve")
+@TargetClass(className = "coursier.exec.Execve")
 @Platforms({Platform.LINUX.class, Platform.DARWIN.class})
 final class ExecveGraalvm {
