+++ modules/exec/src/main/java/coursier/exec/GraalvmUnistdExtras.java
@@ -1,4 +1,4 @@
-package coursier.jvm;
+package coursier.exec;
 
 import com.oracle.svm.core.posix.headers.PosixDirectives;
 import org.graalvm.nativeimage.Platform;
@@ -8,7 +8,6 @@ import org.graalvm.nativeimage.c.function.CFunction;
 import org.graalvm.nativeimage.c.type.CCharPointer;
 import org.graalvm.nativeimage.c.type.CCharPointerPointer;
 
-@Deprecated
 @CContext(PosixDirectives.class)
 @Platforms({Platform.LINUX.class, Platform.DARWIN.class})
 public class GraalvmUnistdExtras {
